MVVM-C
Simple iOS app to showcase the use of the Coordinator pattern using SwiftUI and MVVM.

The implementation is as easy as calling a push and pop methods on the Coordinators. This approach deallocates automatically the view model and coordinator instances when they are popped from the navigation stack and the Views are only re-rendered when needed.